Make "sandbox" a WABT feature.
Instead of having an ad-hoc option just in wasm2c, this makes it
possible to specify sandbox or no-sandbox mode across all wabt tools.
(Most tools are currently not sensitive to it.)

This changes how the option is given on the command line.
Instead of writing --no-sandbox, now write --disable-sandbox.