Avoid all-zero initializers.
When generating initializers for data bytes, checks
whether all bytes in question are zero.  If that is
the case, then we generate an empty initializer list,
relying on default initialization to 0.
(That way a WASM data segment such as BSS that is all zero will
eventually again wind up in the (native) BSS segment.)