Generate proper main and include open proxies.
In no-sandbox mode, generate a "proper" main function when
__main_argc_argv is being exported.  This "proper" main
has the correct types and just calls the other with appropriate casts.

Also include definitions for WASM special functions wasm_open2 and
wasm_open3.  Arrange for any "special" functions (starting with "wasm_")
to not be considered imports.

Finally, move the declarations of data imports into the .c file.
(There is no need for them to be in the .h file.)